AesCtrDecryptor: Move TEE_crypto_init() call into AesCtrDecryptor object.
TEE_crypto_init() allocates memory for IV and key currently.
It is linked to the lifetime of the AES decryptor object
so makes no sense to be called from SessionLibrary object.

Also this change allows the AesCtrDecryptorUnitTests to be run
without any modification making maintaiing the OPTEE ClearKey
and rebasing to new versions of AOSP easier.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>